requirements.txt
cffi==1.9.1
coverage==4.3.1
cryptography==1.7.1
decorator==4.0.11
Django==1.10.4
django-cors-headers==1.3.1
django-filter==1.0.1
django-nose==1.4.4
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.1
enum34==1.1.6
GDAL==1.11.3
gunicorn==19.6.0
idna==2.2
inflection==0.3.1
ipaddress==1.0.17
jmespath==0.9.0
jsonpath-rw==1.4.0
Markdown==2.6.7
meld3==1.0.2
nose==1.3.7
ply==3.10
pyasn1==0.1.9
pycparser==2.17
pygobject==3.18.2
PyJWT==1.4.2
six==1.10.0
supervisor==3.1.3
vboxapi==1.0
wheel==0.26.0
whitenoise==3.2.2
